Apply labels to WEPs using CNI_ARGS
The intention of this commit is to allow any runtime to apply labels to
workload endpoints using `CNI_ARGS` without requiring runtime-specific
handling.

If you supply the following `CNI_ARGS` string to the plugin:

    CALICO_role=prod;CALICO_app=fe;secret=opensesame

The following labels will be applied to the workload endpoint:

    role:prod
    app:fe

Note that only the keys prefixed with `CALICO_` are applied
(`secret:opensesame` is not applied), after stripping the `CALICO_`
prefix.

The `parseArgs` function is copied almost verbatim from:

https://github.com/containernetworking/cni/blob/3805b13a3b5d4fd45187f2572557e5b0e0b1b686/plugins/test/noop/main.go#L56